Proj-JDRES-Clean/requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3